Don't cleanup projects when importing helper
